flask/pyproject.toml
[tool.poetry]
name = "flask-weather"
version = "0.1.0"
description = ""
authors = ["hsuanchi <a0025071@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.2"
uWSGI = "^2.0.19"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
